+++ Changelog.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Changed
+- Remove dead code in `SmartyRenderer::setHelperPluginManager()`
 
 ## [1.0.0] - 2019-08-14
 ### Added